requirements-dev.txt
pytest==3.7.1
pytest-django==3.3.3
flake8==3.5.0
mock==2.0.0
responses==0.9.0
